Addressed issues noted by Sam and Carmen.
