+++ client/src/components/AddProduct.js
@@ -13,134 +13,9 @@ import {
   Select,
 } from "@material-ui/core";
 import CloseIcon from "@material-ui/icons/Close";
+import useStyles from "../styles/AddProductStyles";
 
-const useStyles = makeStyles((theme) => ({
-  "@global": {
-    "*::-webkit-scrollbar": {
-      width: "0.2rem",
-    },
-    "*::-webkit-scrollbar-track": {
-      "-webkit-box-shadow": "inset 0 0 6px rgba(10,10,10,0.1)",
-    },
-    "*::-webkit-scrollbar-thumb": {
-      backgroundColor: "rgba(1,1,1)",
-      outline: "1px solid slategrey",
-    },
-  },
-  popup: {
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  paper: {
-    position: "absolute",
-    display: "flex",
-    flexDirection: "column",
-    alignItems: "center",
-    backgroundColor: "#f8f8f8",
-    boxShadow: theme.shadows[5],
-    padding: theme.spacing(2, 4, 3),
-    borderRadius: "0.3rem",
-    overflow: "hidden",
-    outline: "none",
-    minWidth: "30rem",
-    maxHeight: "41rem",
-    height: "75%",
-  },
-  addButton: {
-    borderRadius: "10rem",
-    padding: theme.spacing(2, 6),
-    backgroundColor: "#DF1B1B",
-    color: "white",
-    margin: theme.spacing(2),
-    width: "11rem",
-    height: "3.5rem",
-    fontSize: "0.8rem",
-  },
-  titleContainer: {
-    display: "flex",
-    maxHeight: "3rem",
-  },
-  closeButtonContainer: {
-    textAlign: "right",
-    width: "100%",
-  },
-  titleText: {
-    fontSize: "1.5rem",
-    fontWeight: "bold",
-  },
-  itemText: {
-    fontSize: "0.8rem",
-    fontWeight: "400",
-    letterSpacing: "0.5px",
-  },
-  productCard: {
-    display: "flex",
-    flexDirection: "row",
-    maxWidth: "30rem",
-  },
-  removeButtonContainer: {
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  removeButton: {
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-    borderRadius: "10rem",
-    height: "3rem",
-  },
-  addButtonContainer: {
-    display: "flex",
-    justifyContent: "center",
-    margin: theme.spacing(3, 0, 2),
-  },
-  bodyContainer: {
-    minWidth: "30rem",
-    maxHeight: "29rem",
-    overflow: "hidden",
-    display: "flex",
-    justifyContent: "center",
-  },
-  bodyContent: {
-    minWidth: "30rem",
-    maxHeight: "30rem",
-    overflow: "scroll",
-    overflowX: "hidden",
-  },
-  textFieldContainer: {
-    display: "flex",
-    justifyContent: "center",
-    width: "80%",
-    backgroundColor: "white",
-  },
-  centerText: {
-    textAlign: "center",
-  },
-  imageFieldContainer: {
-    display: "flex",
-    backgroundColor: "white",
-    width: "35%",
-    height: "0",
-    paddingBottom: "35%",
-  },
-  selectListInput: {
-    display: "flex",
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "center",
-    backgroundColor: "white",
-    margin: "3rem",
-    overflow: "hidden",
-    height: "4rem",
-    padding: "0 0.5rem",
-    boxShadow: "0 0 0.5rem 1px #eee",
-    width: "80%",
-  },
-}));
-
-const AddList = (props) => {
+const AddProduct = (props) => {
   const {
     listTitle,
     listOpen,
@@ -150,7 +25,7 @@ const AddList = (props) => {
     demoListsArray,
   } = props;
   const classes = useStyles();
-  const [title, setTitle] = useState("");
+  const [url, setUrl] = useState("");
   const [list, setList] = useState("");
   const [errors, setErrors] = useState({});
 
@@ -164,7 +39,7 @@ const AddList = (props) => {
 
   const validations = () => {
     const errorsCopy = { ...errors };
-    errorsCopy.title = title ? "" : "This field is required.";
+    errorsCopy.url = url ? "" : "This field is required.";
     setErrors({ ...errorsCopy });
 
     return Object.values(errorsCopy).every((field) => field === "");
@@ -179,7 +54,7 @@ const AddList = (props) => {
           "Content-Type": "application/json",
         },
         body: JSON.stringify({
-          title: title,
+          url: url,
           //'cover': cover
         }),
       })
@@ -232,9 +107,9 @@ const AddList = (props) => {
             placeholder="Paste your link here" // Placeholder needs to be centered.
             fullWidth
             type="text"
-            error={!!errors.title}
-            helperText={errors.title}
-            onChange={(e) => setTitle(e.target.value)}
+            error={Boolean(errors.url)}
+            helperText={errors.url}
+            onChange={(e) => setUrl(e.target.value)}
           />
         </Box>
         <Box className={classes.bodyContainer}>
@@ -271,4 +146,4 @@ const AddList = (props) => {
   );
 };
 
-export default AddList;
+export default AddProduct;
