+++ client/src/pages/ShoppingLists.js
@@ -2,15 +2,20 @@ import React, { useState } from "react";
 import { Box } from "@material-ui/core";
 import AddItem from "components/AddItem";
 import ListsDisplay from "components/ListsDisplay";
-import AddList from "components/AddList"
+import AddList from "components/AddList";
 
 const ShoppingLists = () => {
-    const [addListOpen, setAddListOpen] = useState(false)
+  const [addListOpen, setAddListOpen] = useState(false);
+
+  const changeAddListOpen = () => {
+    setAddListOpen((previous) => !previous);
+  };
+
   return (
     <Box>
       <AddItem />
-      <ListsDisplay addListOpen = {addListOpen} setAddListOpen = {setAddListOpen} />
-      <AddList addListOpen = {addListOpen} setAddListOpen = {setAddListOpen}/>
+      <ListsDisplay {...{ addListOpen, changeAddListOpen }} />
+      <AddList {...{ addListOpen, changeAddListOpen }} />
     </Box>
   );
 };
